refactor: 타입 추가 분리 및 DropdownField 컴포넌트 추가
- submission.ts에서 기존 SubmissionResponse에서 SubmissionResult를 추가분리
- 데이터 매핑 로직인 mapSubmission도 type/submission.ts로 분리
- useMemo를 사용해 데이터를 Submission[] 형태로 매핑 처리
